pyproject.toml
[tool.poetry]
name = "whisper-speech-recogniniton"
version = "1.1.0"
description = "This project is a transcription service built using FastAPI. It transcribes audio files using the WhisperTranscriber model."
authors = ["Vladimir Katin <katin.v.v@gmail.com>"]
license = "GNU GPL v3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai-whisper = "^20231117"
minio = "7.2.5"
fastapi = "0.110.0"
uvicorn = "0.29.0"
pydantic = "2.6.4"
python-dotenv = "1.0.1"
llvmlite = "^0.42.0"
numba = "^0.59.1"
opentelemetry-api = "^1.23.0"
opentelemetry-sdk = "^1.23.0"
torch = { version = "2.2.1", source = "torch" }
opentelemetry-exporter-otlp-proto-grpc = "^1.23.0"
opentelemetry-instrumentation-fastapi = "^0.44b0"
aiokafka = "^0.10.0"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.group.test.dependencies]
flake8 = "^7.0.0"
black = "^24.2.0"
isort = "^5.13.2"
pytest = "^8.1.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
torch = "2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
